backend/services/GraphQL_Service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
strawberry-graphql[fastapi]==0.252.3
httpx==0.28.1
python-dotenv==1.0.1
